more cleanup and refactoring / seperating files
