+++ lib/toolbox.js
@@ -1,0 +1,18 @@
+module.exports = `
+  <xml id="toolbox" style="display: none">
+      <category name="Logic">
+        <block type="controls_if"></block>
+        <block type="controls_repeat_ext"></block>
+        <block type="logic_compare"></block>
+        <block type="math_number"></block>
+        <block type="math_arithmetic"></block>
+        <block type="text"></block>
+        <block type="text_print"></block>
+      </category>
+      <sep></sep>
+      <category name="Movement" colour="210">
+        <block type="move"></block>
+        <block type="rotate"></block>
+      </category>
+   </xml>
+`
