+++ src/components/blockly.js
@@ -3,25 +3,7 @@
 const widget = require('cache-element/widget')
 const html = require('choo/html')
 const sf = require('sheetify')
-
-const toolbox = `
-  <xml id="toolbox" style="display: none">
-      <category name="Logic">
-        <block type="controls_if"></block>
-        <block type="controls_repeat_ext"></block>
-        <block type="logic_compare"></block>
-        <block type="math_number"></block>
-        <block type="math_arithmetic"></block>
-        <block type="text"></block>
-        <block type="text_print"></block>
-      </category>
-      <sep></sep>
-      <category name="Movement" colour="210">
-        <block type="move"></block>
-        <block type="rotate"></block>
-      </category>
-   </xml>
-`
+const toolbox = require('../../lib/toolbox')
 
 const prefix = sf`
   :host, :host > .editor {
